refactor(agents): relocate AgentTraces under observability/traces [PLT-103787]
Move the agent trace-level analytics service from agents/traces into
observability/traces/agent (files renamed traces.* → agent.*), placing it
beside the existing observability Traces service. Consolidate its public
surface into the /traces subpath (export AgentTraces from
observability/traces/index.ts) and drop the dedicated ./agent-traces subpath
(package.json + rollup) — it was unreleased, so no consumer impact. Symbols
(AgentTracesService, AgentTraces, SpanResponse, AgentTrace* types) unchanged;
no collisions with the observability traces types. Tests mirrored to the new
location; unified-setup + mkdocs nav updated. AGENT_TRACES_ENDPOINTS kept.
