--- src/models/agents/traces/index.ts
@@ -1,2 +1,0 @@
-export * from './traces.types';
-export * from './traces.models';
