+++ src/models/observability/traces/agent/index.ts
@@ -1,0 +1,2 @@
+export * from './agent.types';
+export * from './agent.models';
