+++ src/services/observability/traces/index.ts
@@ -1,2 +1,4 @@
 export { TracesService as Traces } from './traces';
 export * from '../../../models/observability/traces';
+
+export * from './agent';
